+++ src/types/PaymentIntent.ts
@@ -51,7 +51,8 @@ export type ConfirmParams =
   | PayPalParams
   | AffirmParams
   | CashAppParams
-  | RevolutPayParams;
+  | RevolutPayParams
+  | PayByBankParams;
 
 export type ConfirmOptions = PaymentMethod.ConfirmOptions;
 
@@ -309,6 +310,15 @@ export type CashAppParams = {
 
 export type RevolutPayParams = {
   paymentMethodType: 'RevolutPay';
+  paymentMethodData?: {
+    billingDetails?: BillingDetails;
+    mandateData?: MandateData;
+    metadata?: MetaData;
+  };
+};
+
+export type PayByBankParams = {
+  paymentMethodType: 'PayByBank';
   paymentMethodData?: {
     billingDetails?: BillingDetails;
     mandateData?: MandateData;